componentize service status 💚🔴
