+++ components/ServiceRow/ServiceRow.js
@@ -1,21 +1,7 @@
 import React, { PropTypes } from 'react'
 import './ServiceRow.scss'
 
-const isOpen = (schedules) => {
-  const currentDay = 'Wednesday'
-  const currentTime = 1400
-
-  const isBetween = (num, min, max) => (
-    num > min && num < max
-  )
-
-  const isSomethingSchedule = schedule => (
-    currentDay === schedule.weekday &&
-    isBetween(currentTime, schedule.opensAt, schedule.closesAt)
-  )
-
-  return schedules.filter(isSomethingSchedule).length > 0
-}
+import ServiceStatus from '../ServiceStatus'
 
 const ServiceRow = (props) => (
   <li className="service">
@@ -25,7 +11,7 @@ const ServiceRow = (props) => (
             <p className="serviceName">{props.name}</p>
           </div>
             <span className="serviceInfo">
-              <span className="label-status something">{isOpen(props.regularSchedules) ? 'open' : 'closed'}</span>
+              <ServiceStatus schedules={props.regularSchedules} />
             </span>
           <div className="categoryIcons">
               <span><i className="category-icon icon-home"></i></span>
