+++ components/ServiceStatus/ServiceStatus.scss
@@ -1,0 +1,7 @@
+.open {
+
+}
+
+.closed {
+
+}
